one-hot

+++ src/single_image_model.py
@@ -426,7 +426,9 @@ class SingleImageModel:
 
 		if self.n_features > 0:
 
-			xf_onehot = tf.one_hot(self.xf, self.dataloader.num_pids)
+			xf_onehot = tf.one_hot(
+				tf.cast(self.xf, dtype=tf.int32),
+				self.dataloader.num_pids)
 			all_features = tf.concat([self.image_features, xf_onehot], axis=1)
 
 		else:
